0033 - Refaktorisanje modela

+++ ISA-MRS-project/go-fishing-api/src/main/java/tim7/ISAMRSproject/repository/CottageRepository.java
@@ -12,6 +12,6 @@ import tim7.ISAMRSproject.model.Cottage;
 public interface CottageRepository extends JpaRepository<Cottage,Integer>{
 	
 	@Query("select v from Cottage v where v.cottageOwner.id = ?1")
-	public List<Cottage> findByOwnerId(Long ownerId);
+	public List<Cottage> findByOwnerId(Integer ownerId);
 	
 }
